Hide implementation details again
